requirements.txt
langchain
langchain-classic
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
sentence-transformers
streamlit
python-dotenv
pypdf
beautifulsoup4
lxml